+++ weblate/settings_docker.py
@@ -881,6 +881,7 @@ CHECK_LIST = [
     "weblate.checks.source.OptionalPluralCheck",
     "weblate.checks.source.EllipsisCheck",
     "weblate.checks.source.MultipleFailingCheck",
+    "weblate.checks.source.LongUntranslatedCheck",
     "weblate.checks.format.MultipleUnnamedFormatsCheck",
 ]
 modify_env_list(CHECK_LIST, "CHECK")